DEV: Rename `theme-rollup` -> `asset-processor-rollup`
This file has been used for theme compilation for some time. Rename to avoid confusing humans or robots
